Implement reacting

+++ app/controllers/admin/reacts_controller.rb
@@ -4,11 +4,15 @@ class Admin::ReactsController < Admin::BaseAdminController
   before_action :set_timeline
 
   def toggle
+    action = params[:react]
     react = @timeline.reacts.where(reactor: current_active_user).first
-    if react
+
+    if react && react.react == action
       react.destroy
+    elsif react && react.react != action
+      react.update(react: action)
     else
-      React.create(timeline: @timeline, reactor: current_active_user)
+      React.create(timeline: @timeline, reactor: current_active_user, react: action)
     end
   end
